Output filenames and a management system
When we output the JSON text file, the filename is identical for all
figures. Due to we may have the requirement of doing the parallel
computing, we decide to make the figures mapped to different JSON text
file now. Then, an alpha release of a management system
is uploaded. We continually update it too.